Refactor mock exam flow and improve leaderboard UI
Moved the assessment logic from the mock exam details page to a dedicated assessment page for better separation of concerns. Enhanced the leaderboard UI with avatars and user initials, and updated backend leaderboard data to include full user objects. Improved exam details display and added a utility for extracting user initials. Updated navigation logic to direct users to the exam details page after completion.
