+++ app/(admin)/admin/exams/[id]/_components/ExamDetails.tsx
@@ -4,9 +4,7 @@ import { useQuery } from "convex/react"
 import { format } from "date-fns"
 import { fr } from "date-fns/locale"
 import { Calendar, Clock, FileText } from "lucide-react"
-import { useState } from "react"
 import ExamStatusBadge from "@/components/admin/exam-status-badge"
-import { Button } from "@/components/ui/button"
 import {
   Card,
   CardContent,
@@ -14,14 +12,6 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card"
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog"
 import { api } from "@/convex/_generated/api"
 import { Id } from "@/convex/_generated/dataModel"
 import { getExamStatus } from "@/lib/exam-status"
@@ -31,7 +21,6 @@ import { ExamSectionStats } from "./exam-section-stats"
 
 export function ExamDetails({ examId }: { examId: Id<"exams"> }) {
   const exam = useQuery(api.exams.getExamWithQuestions, { examId })
-  const [isDescriptionOpen, setIsDescriptionOpen] = useState(false)
 
   if (!exam) {
     return (
@@ -49,38 +38,16 @@ export function ExamDetails({ examId }: { examId: Id<"exams"> }) {
       {/* En-tête de l'examen */}
       <Card>
         <CardHeader>
-          <div className="flex flex-col gap-2">
-            <div className="flex items-center justify-between">
+          <div className="flex items-start justify-between">
+            <div className="flex flex-col gap-1">
               <CardTitle className="text-2xl text-blue-600 dark:text-white">
                 {exam.title}
               </CardTitle>
-              <ExamStatusBadge status={status} />
+              {exam.description && (
+                <CardDescription>{exam.description}</CardDescription>
+              )}
             </div>
-            {exam.description && (
-              <div className="mt-2">
-                <Dialog
-                  open={isDescriptionOpen}
-                  onOpenChange={setIsDescriptionOpen}
-                >
-                  <DialogTrigger asChild>
-                    <Button size="sm" variant="outline">
-                      Afficher la description
-                    </Button>
-                  </DialogTrigger>
-                  <DialogContent className="dark:bg-card bg-white sm:max-w-[500px]">
-                    <DialogHeader>
-                      <DialogTitle className="text-blue-600 dark:text-white">
-                        Description de l&apos;examen {exam.title}
-                      </DialogTitle>
-                      <DialogDescription>
-                        Détails et informations sur cet examen.
-                      </DialogDescription>
-                    </DialogHeader>
-                    <CardDescription>{exam.description}</CardDescription>
-                  </DialogContent>
-                </Dialog>
-              </div>
-            )}
+            <ExamStatusBadge status={status} />
           </div>
         </CardHeader>
         <CardContent>
