+++ app/(admin)/admin/exams/[id]/_components/ExamLeaderboard.tsx
@@ -3,6 +3,7 @@
 import { useQuery } from "convex/react"
 import { format } from "date-fns"
 import { fr } from "date-fns/locale"
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import {
   Card,
   CardContent,
@@ -13,6 +14,7 @@ import {
 import { Input } from "@/components/ui/input"
 import { api } from "@/convex/_generated/api"
 import { Id } from "@/convex/_generated/dataModel"
+import { getInitials } from "@/lib/utils"
 
 export function ExamLeaderboard({ examId }: { examId: Id<"exams"> }) {
   const leaderboard = useQuery(api.exams.getExamLeaderboard, { examId })
@@ -20,7 +22,7 @@ export function ExamLeaderboard({ examId }: { examId: Id<"exams"> }) {
   if (!leaderboard || leaderboard.length === 0) return null
 
   return (
-    <Card>
+    <Card className="@container">
       <CardHeader>
         <div className="flex flex-col gap-3 md:flex-row md:items-center md:justify-between">
           <div>
@@ -39,31 +41,40 @@ export function ExamLeaderboard({ examId }: { examId: Id<"exams"> }) {
       </CardHeader>
       <CardContent>
         <div className="space-y-4">
-          {leaderboard.map((entry, index) => (
-            <div key={index} className="flex items-center justify-between">
-              <div className="flex items-center gap-3">
-                <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-blue-100 text-sm font-medium text-blue-600 dark:bg-blue-900 dark:text-blue-400">
-                  {index + 1}
+          {leaderboard.map((entry, index) => {
+            const initials = getInitials(entry.user?.name)
+            return (
+              <div
+                key={index}
+                className="flex items-center justify-between rounded-lg border p-2"
+              >
+                <div className="flex min-w-0 flex-1 items-center gap-3">
+                  <div className="flex h-6 w-6 shrink-0 items-center justify-center rounded bg-blue-600 text-xs font-bold text-white dark:bg-blue-500">
+                    {index + 1}
+                  </div>
+                  <Avatar className="size-10 shrink-0">
+                    <AvatarImage
+                      src={entry.user?.image}
+                      alt={entry.user?.name || "Avatar"}
+                    />
+                    <AvatarFallback>{initials}</AvatarFallback>
+                  </Avatar>
+                  <div className="min-w-0 flex-1">
+                    <p className="truncate font-medium">{entry.user?.name}</p>
+                    <p className="text-muted-foreground truncate text-sm">
+                      @{entry.user?.username}
+                    </p>
+                  </div>
                 </div>
-                <div>
-                  <p className="font-medium">
-                    {entry.user?.username && (
-                      <span className="text-muted-foreground ml-2 text-sm">
-                        @{entry.user.username}
-                      </span>
-                    )}
-                    {entry.user?.name}
+                <div className="shrink-0 text-right">
+                  <p className="font-bold">{entry.score}%</p>
+                  <p className="text-muted-foreground hidden text-sm @md:block">
+                    {format(new Date(entry.completedAt), "Pp", { locale: fr })}
                   </p>
                 </div>
               </div>
-              <div className="text-right">
-                <p className="font-bold">{entry.score}%</p>
-                <p className="text-muted-foreground text-sm">
-                  {format(new Date(entry.completedAt), "PPP", { locale: fr })}
-                </p>
-              </div>
-            </div>
-          ))}
+            )
+          })}
         </div>
       </CardContent>
     </Card>
